Refactor gallery logic into single module
Merged works and filters functionality into a new gallery.js module, replacing getWorks.js and getFilters.js. Updated main.js to use the unified gallery module for initializing and rendering works and filters, improving maintainability and reducing code duplication.
