+++ FrontEnd/assets/script/main.js
@@ -1,12 +1,12 @@
-import  {getWorksInit} from "./modules/getWorks.js";
-import  {getFiltersInit} from "./modules/getFilters.js";
-
+import  {getWorksInit,getFiltersInit} from "./modules/gallery.js";
 
 /**
- * Initializes the application by fetching and displaying works.
+ * Initializes the application by fetching and rendering works and filters.
+ * Calls getWorksInit to display works in the gallery and getFiltersInit to set up filter buttons.
+ *
  * @async
  * @function init
- * @returns {Promise<void>}
+ * @returns {Promise<void>} Resolves when initialization is complete.
  */
 async function init() {
     getWorksInit('works', '.gallery');
